SuperUser Functionality Added
Delete your db and sync again and user the Superuser that you create here for the superuser account in the portal as well.
The Superuser can create groups with multiple permissions.(NOTE: The permissions haven't been customized yet and so you can
 select any permissions while creating groups). Once a group has been created, selected the corresponding edit button to add cores
to that group. The cores added here can use the login feature to goto cores_home.
NOTE: If you have logged into a coord or a core account and the respective home page doesn't have a logout button still,
you can directly go to http://127.0.0.1:8000/logout to return back to home.